Add convex hull query
